+++ src/common/config_parser.cpp
@@ -398,9 +398,10 @@ void ConfigParser::addOptionsTraining(cli::CLIWrapper& cli) {
   cli.add<std::string/*SchedulerPeriod*/>("--save-freq",
       "Save model file every  arg  updates (append 't' for every  arg  target labels)",
       "10000u");
-  cli.add<std::string>("--logical-epoch",
-      "Redefine logical epoch counter as multiple of data epochs (e.g. 1e), updates (e.g. 100Ku) or labels (e.g. 1Gt)",
-      "1e");
+  cli.add<std::vector<std::string>>("--logical-epoch",
+      "Redefine logical epoch counter as multiple of data epochs (e.g. 1e), updates (e.g. 100Ku) or labels (e.g. 1Gt). "
+      "Second parameter defines width of fractional display, 0 by default.",
+      {"1e", "0"});
 
   addSuboptionsInputLength(cli);
   addSuboptionsTSV(cli);